added function to upload dataframe as csv

+++ S3Connector/S3Connector.py
@@ -1,4 +1,6 @@
+import io
 import os
+import re
 import sys
 
 import boto3
@@ -9,23 +11,30 @@ sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
 class S3Connector(object):
 
-	def __init__(self, aws_access_key_id=None, aws_secret_access_key=None, aws_region_name=None):
-		self._create_clients(aws_access_key_id, aws_secret_access_key, aws_region_name)
+	def __init__(self, aws_access_key_id=None, aws_secret_access_key=None, aws_region=None):
+		self._create_clients(aws_access_key_id, aws_secret_access_key, aws_region)
 
-	def _create_clients(self, aws_access_key_id=None, aws_secret_access_key=None, aws_region_name=None):
+	def _create_clients(self, aws_access_key_id=None, aws_secret_access_key=None, aws_region=None):
 
-		if None not in [aws_access_key_id, aws_secret_access_key, aws_region_name]:
+		if aws_region is None:
+			aws_region = 'eu-central-1'
+
+		self._aws_access_key_id = aws_access_key_id,
+		self._aws_secret_access_key = aws_secret_access_key,
+		self._aws_region = aws_region
+
+		if None not in [self._aws_access_key_id, self._aws_secret_access_key, self._aws_region]:
 			self._s3_client = boto3.client(
 				's3',
-				aws_access_key_id=aws_access_key_id,
-				aws_secret_access_key=aws_secret_access_key,
-				region_name=aws_region_name,
+				aws_access_key_id=self._aws_access_key_id,
+				aws_secret_access_key=self._aws_secret_access_key,
+				region_name=self._aws_region,
 			)
 			self._s3_resource = boto3.resource(
 				's3',
-				aws_access_key_id=aws_access_key_id,
-				aws_secret_access_key=aws_secret_access_key,
-				region_name=aws_region_name,
+				aws_access_key_id=self._aws_access_key_id,
+				aws_secret_access_key=self._aws_secret_access_key,
+				region_name=self._aws_region,
 			)
 		self._s3_client = boto3.client('s3')
 		self._s3_resource = boto3.resource('s3')
@@ -42,6 +51,7 @@ class S3Connector(object):
 		:param bucket_name: string
 		:return: True if the referenced bucket_name exists, otherwise False
 		"""
+		if bucket_name is None: return
 		try:
 			resp = self._s3_client.head_bucket(Bucket=bucket_name)
 		except botocore.exceptions.ClientError as e:
@@ -50,6 +60,8 @@ class S3Connector(object):
 		return True
 
 	def _create_bucket(self, bucket_name=None):
+		if bucket_name is None: return
+
 		if self._exists_bucket(bucket_name):
 			print("Cannot create the bucket. A bucket with the name '" +
 				  bucket_name + "' already exists. Exiting.")
@@ -67,6 +79,8 @@ class S3Connector(object):
 		return
 
 	def _delete_bucket(self, bucket_name):
+		if bucket_name is None: return
+
 		print("Deleting the bucket named '" + bucket_name + "'...")
 		self._s3_client.delete_bucket(Bucket=bucket_name)
 		print("Bucket deleted")
@@ -79,6 +93,8 @@ class S3Connector(object):
 		:param object_name: S3 object name. If not specified then basename of file path
 		:return: True if file was uploaded, else False
 		"""
+		if bucket_name is None: return
+
 		# If S3 object_name was not specified, use basename of file_path
 		if object_name is None:
 			object_name = os.path.basename(file_path)
@@ -95,6 +111,44 @@ class S3Connector(object):
 			return False
 		return True
 
+	def upload_dataframe_as_csv(self, dataframe=None, bucket_name=None, object_name=None):
+		"""
+		Upload dataframe as csv to an S3 bucket
+		:param dataframe: pandas dataframe
+		:param bucket_name: Bucket to upload to
+		:param object_name: S3 object name
+		:return: True if file was uploaded, else False
+		"""
+		if dataframe is None: return
+		if bucket_name is None: return
+		if object_name is None: return
+
+		try:
+			# general csv features
+			csv_sep = '\t'
+			csv_null_identifier = '#N/A'
+
+			# save dataframe as temp csv
+			csv_io = io.StringIO()
+			dataframe.to_csv(csv_io, sep=csv_sep, encoding='utf-8-sig',
+							 header=True, index=False, na_rep=csv_null_identifier)
+			csv_contents = csv_io.getvalue()
+			csv_contents = re.sub(r'NaT', csv_null_identifier, csv_contents)
+			csv_io.seek(0)
+			csv_io.write(csv_contents)
+
+			# copy temp csv file to S3 object
+			if not self._exists_bucket(bucket_name):
+				self._create_bucket(bucket_name)
+			self._s3_resource.Object(bucket_name, object_name).put(Body=csv_io.getvalue())
+			csv_io.close()
+
+			print('File uploaded')
+		except Exception as e:
+			print(e)
+			return False
+		return True
+
 	def download_file(self, bucket_name=None, object_name=None, file_path=None):
 		"""
 		Download a file from S3 bucket
@@ -103,6 +157,9 @@ class S3Connector(object):
 		:param file_path: str
 		:return: True if the file was download, else False
 		"""
+		if bucket_name is None: return
+		if object_name is None: return
+
 		try:
 			self._s3_resource.Object(bucket_name, object_name).download_file(file_path)
 		except Exception as e:
